+++ src/protocol/nefarious13.c
@@ -1407,7 +1407,7 @@ void nefarious_cmd_ping(char *server)
 	send_cmd(p10id, "RI %s %s %ld %ld :<No client start time>",
 	         ((s
 	           && s->suid) ? s->suid : server),
-	         (ud ? ud->uid : s_StatServ), t.tv_sec, t.tv_usec);
+	         (ud ? ud->uid : s_StatServ), (long int) t.tv_sec, (long int) t.tv_usec);
 }
 
 void nefarious_cmd_ctcp(char *source, char *dest, char *buf)
