+++ src/protocol/scarynet.c
@@ -1163,7 +1163,7 @@ void scarynet_cmd_ping(char *server)
 	gettimeofday(&t, NULL);
 	send_cmd(p10id, "RI %s %s %ld %ld :<No client start time>",
 	         ((s && s->suid) ? s->suid : server),
-	         (ud ? ud->uid : s_StatServ), t.tv_sec, t.tv_usec);
+	         (ud ? ud->uid : s_StatServ), (long int) t.tv_sec, (long int) t.tv_usec);
 }
 
 void scarynet_cmd_ctcp(char *source, char *dest, char *buf)
